EECE5698-Reinforcement-Learning/maze_navigation/requirements.txt
torch>=2.0.0
gymnasium>=0.29.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
